backend/requirements.txt
fastapi
uvicorn
paho-mqtt
SQLAlchemy
pydantic
psycopg2-binary
